tmkms-p2p: flatten crate hierarchy
Removes the second level of hierarchy under the `secret_connection`
module, flattening out the crate to a single level, and extracting more
types from `secret_connection` into their own submodules where
appropriate.

This also removes quite a bit from the crate's public API which doesn't
actually need to be exposed.
