+++ tmkms-p2p/src/lib.rs
@@ -1,7 +1,10 @@
 //! The Tendermint P2P stack.
 
 #![forbid(unsafe_code)]
-#![deny(
+#![warn(
+    clippy::all,
+    clippy::pedantic,
+    clippy::unwrap_used,
     nonstandard_style,
     rust_2018_idioms,
     trivial_casts,
@@ -9,21 +12,28 @@
     unused_import_braces,
     unused_qualifications
 )]
-#![warn(
-    clippy::all,
-    clippy::nursery,
-    clippy::pedantic,
-    clippy::unwrap_used,
-    unused_import_braces,
-    unused_qualifications
-)]
 #![doc(
     html_logo_url = "https://raw.githubusercontent.com/informalsystems/tendermint-rs/master/img/logo-tendermint-rs_3961x4001.png"
 )]
 
-pub mod secret_connection;
 pub mod transport;
 
 mod error;
+mod handshake;
+mod kdf;
+mod nonce;
+mod protocol;
+mod public_key;
+mod secret_connection;
+mod state;
 
-pub use error::{Error, Result};
+pub use crate::{
+    error::{Error, Result},
+    public_key::PublicKey,
+    secret_connection::SecretConnection,
+};
+
+pub(crate) use tendermint_proto::v0_38 as proto;
+
+/// Maximum size of a message
+pub const DATA_MAX_SIZE: usize = 1024;
